Support bulk registration of people.
Similarly to the support for bulk registration of countries, support
is added for registering participants (staff only) in bulk by
uploading a CSV file.  Only limited data can be registered this way,
with the rest to be added by the participants in question through the
self-registration mechanism; if contact email addresses are given in
the CSV file, self-registration accounts are automatically created and
email with password details sent to those addresses.