Refactor registration system locking code out of caching code.
Factoring this out into a separate function and file has two benefits.

(a) While the present caching of scoreboard text requires locking to
work correctly, so the scoreboard must be regenerated every time if
locking is unsupported to avoid an out-of-date copy possibly being
cached and considered current, there are other uses for which caching
is OK even without locking.  In particular, it is useful to generate
thumbnail images of flags and photographs on the server, and because
any change to those involves creating a new flag or photo item rather
than changing the contents of an existing one, such thumbnails can
readily be cached even without locking (but with the possibility in
the absence of locking that multiple processes might generate the same
cached image at the same time).  By having a factored-out locking
function that can also run when locking is not supported, this use
case is facilitated.

(b) Moving the fcntl-using code to a separate small file avoids
warnings from code checkers about other imports coming after the check
for the availability of the fcntl module.