Factor out some image-scaling code to new matholymp.images.
Further preparation for generating thumbnail images instead of always
sending full-size images to the browser to be scaled there.  Such
thumbnail generation is relevant to both the static site and the
registration system, so image scaling code should be located outside
the registration system code to be used in both places.