Support uploading photos in bulk registration.
Bulk registration of people now supports uploading a ZIP file and
specifying in the CSV file paths in the ZIP file for photos.  Much of
the code is generic and so could also be used if desired to support
uploading flags when bulk registering countries and consent forms when
bulk registering people.