Generate thumbnails for flags on static site.
This avoids the browser needing to load a full-size flag image and
then scale it down client-side.  This is in preparation for doing the
same for photos, both on the static site and the registration system,
where the amount of data involved in loading full-size photos
unnecessarily is much more significant, especially when loading the
country page for a staff country with many participants and so many
photos on the page.