Refactor handling of comma-separated configuration variables for registration system.
There are seven configuration variables in the registration system
with comma-separated values, all following the same rules that leading
and trailing whitespace is removed from each item in the list and
items that are empty after that are ignored.  This moves that logic
into a common function get_config_var_comma_sep, with one function per
variable in matholymp.roundupreg.config that uses it to get the value
of the variable.