Support generating papers for all contestants in separate files.
Support for generating a paper for an individual contestant was
already present, by specifying a contestant code on the
mo-document-generate command line.  This adds support for generating
papers for all contestants in a single run of mo-document-generate, by
using "all-split" in place of "all".  This option, intended for a
virtual event, does not generate any files of papers for leaders and
deputies.