Do not collect irrelevant information for a virtual event.
Information about allergies and dietary requirements, travel, room
type, room number and person to share a room with is definitely
irrelevant for a virtual event, so do not show or collect that
information in person.item.html in that case.  Allow dietary
requirements, consent information for dietary requirements and room
type to be missing and avoid default room type in the case of a
virtual event, even when that information would otherwise be required
or defaulted.

Some other information, such as name as on passport, is probably also
irrelevant, but the user interface for it is left unchanged, should
the relevant configuration options be set (which they probably
wouldn't be for an event that starts out as virtual).