Do not collect passport / nationality details for a virtual event.
Hide the user interface for collection nationality, passport number
and name as on passport in the case of a virtual event and do not
treat any of rhat information as required.

Since a virtual event can simply disable the relevant configuration
options, the main purpose of this change is as preparation for
supporting hybrid events, where some participants take part in person
and others remotely and where whether this information is required for
a participant depends on whether they are taking part in person.