Support hybrid events in data model and static site generation.
Add support in the data model for hybrid events (some participants
taking part in person and others remotely), in addition to in-person
and virtual events, with corresponding support in static site
generation.  This only covers the change to the data model for events;
tracking in-person state for people and countries will be handled
separately, and corresponding changes to the generation of the
registration status page will come with the changes for people.
Release notes to follow with registration system support.