Track remote participation of individual participants in data model.
Another piece of preparation for supporting hybrid events.  Whether an
individual is participating remotely is tracked in the data model and
in the full CSV file of participants and this controls the relevant
parts of the registration status page (missing information on remote
participation is listed as needed, but travel details and room number
aren't for remote participants).  Actual registration system support
and corresponding release notes to be added separately (after adding
similar support in the data model for tracking how countries are
participating).