Support hybrid events in registration system.
Checks for virtual events are updated where appropriate to handle
hybrid events.  Whether countries and people expect to participate in
person or remotely (or a combination, in the case of countries) can be
specified separately for each country and person.  There is no current
check for the choice specified for a person being consistent with that
specified for a country (which is just treated as a provisional
specification of intent).

Appropriate handling of hybrid events for generating name badges and
exam papers is to be added separately.