Sort room types by name in error from auditor.
Testing with Roundup 2.1.0 showed the tests of errors for bad room
types for a role failing because the error message listed the valid
room types in an unexpected order.  Sort the room types by name for
this message so it doesn't depend on the order in which Roundup
returns the room types, and update the tests accordingly.