Handle collecting vaccination status information.
There are three questions for the information collected (vaccine
certificate status, number of doses and whether a dose received after
a threshold date), but the precise wording for those questions is left
to be specified in config.ini given the likelihood that the
appropriate details will vary from event to event.  Each question can
be controlled independently via config.ini.  If the user interface for
asking for consent information is enabled, it seeks consent for
processing any of the vaccination data.