Refactor registration_enabled handling.
In places where the check is for whether the current user can register
participants or edit details of registered participants at the current
time, use a registration_enabled function instead of checking the
registration_enabled property.  This would facilitate separating
enabling of registration for participating countries from enabling
that for self-registering users.