Support hiding non-final scores in registration system
A message about hiding scores may be specified on the event page in
the registration system, and if nonempty then scores will be hidden
(if not final) and that message shown on the scoreboard page.  For
example, this may be used to delay displaying scores until contestants
have returned from an excursion (at which point the message would be
set to empty, so opening access to scores).  Users with the new
ViewScores role (or Admin or Score users) may still view the scores
while they are hidden from public view.