Support using xelatex for document generation
This is used automatically if "xelatex" is mentioned anywhere in the
template file (expected to be in a comment), and helps when
registration data includes characters not supported by default by
pdflatex, being less likely to need custom configuration depending on
the specific characters that happen to be used at a given event.
There is no corresponding support for automatic use of xelatex to
build papers that need it, so such papers still need a PDF to be
generated externally.