Start setup for script scans handling: data model
Add information about scans of scripts (URLs and filenames) to data
model and associated CSV file generation.  Scans are to be stored (for
each contestant) as one PDF file per problem plus one per day for
unsorted scratch paper not in a problem folder.

This is just the initial addition of scans to the data model preparing
the way for adding more substantive support to the registration system
later, and does not include any of that substantive support for
storing and processing scans.